render response as preformatted JSON, not HTML (#100)
HTML tables were super-slow if the response was very large and/or complicated in shape. Also, the value of the formatted table was low since it could not also provide any useful metadata (like response types and the like).

So now the response messages are rendered as plain (but multi-line/indented) JSON into a textarea. This is not only much faster but also more functional, since it is a reasonable use case to want to copy out and use the JSON response.